Extracted service descriptor interface to a separate file.
