+++ src/ServiceDescriptor.ts
@@ -1,0 +1,7 @@
+
+export interface ServiceDescriptor {
+  name?: string;
+  type?: Function;
+  factory?: Function;
+  params?: any[];
+}
