+++ test/unit/Container.spec.ts
@@ -1,7 +1,9 @@
+
 import "es6-shim";
 import "reflect-metadata";
 import {Container} from "../../src/Container";
 import {Service} from "../../src/decorators";
+
 
 describe("Container", function() {
